Extract the Provider module (config/test/naming) into Api/Provider
Seventh module of the #777 folder migration. Moves the provider-management
types into Jellyfin.Plugin.SSO_Auth.Api.Provider:

- ProviderConnectionTester, ProviderMode, ProviderNameValidator,
  ProviderTestResult -> Api/Provider/ (namespace Api.Provider), import added at
  each call site.

Provider may depend on Net and RateLimit (ProviderNameValidator validates URLs
via Net and keys throttles via RateLimit). Registers the Provider -> {Net,
RateLimit} edges on the module DAG rule.

Pure mechanical move, no behavior change. Verified by the full build
(--warnaserror, both TFMs), the complete test suite (both TFMs), the module DAG
conformance rule, and the ABI-floor build.

Refs #777
