+++ SSO-Auth/Api/Provider/ProviderConnectionTester.cs
@@ -8,7 +8,7 @@ using Duende.IdentityModel.OidcClient;
 using Jellyfin.Plugin.SSO_Auth.Config;
 using Microsoft.Extensions.Logging;
 
-namespace Jellyfin.Plugin.SSO_Auth.Api;
+namespace Jellyfin.Plugin.SSO_Auth.Api.Provider;
 
 /// <summary>
 /// Runs an admin-triggered Test-connection probe against a STORED provider configuration (#163) so an
