+++ SSO-Auth/Config/ProviderConfigValidator.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using Jellyfin.Plugin.SSO_Auth.Api;
 using Jellyfin.Plugin.SSO_Auth.Api.Authz;
 using Jellyfin.Plugin.SSO_Auth.Api.Net;
+using Jellyfin.Plugin.SSO_Auth.Api.Provider;
 
 namespace Jellyfin.Plugin.SSO_Auth.Config;
